requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
splunk-sdk
pymongo
sshtunnel==0.4.0
paramiko==2.12.0
mysql-connector-python
urllib3
elasticsearch
python-dotenv
